pango plugin pango_textlayout: do overflow check as 'double'
Coverity observes:

  CID 1681838: (#1 of 1): Result is not floating-point
  (UNINTENDED_INTEGER_DIVISION)
  integer_division: Dividing integer expressions 2147483647 and 1024, and then
  converting the integer quotient to type double. Any remainder, or fractional
  part of the quotient, is ignored.

This condition is checking for overflow in a later calculation which is done in
`double` space. So this overflow check should also be using `double` division.

Coverity: https://scan5.scan.coverity.com/#/project-view/66357/10724?selectedIssue=1681838
Reported-by: Coverity 2024.12.1
Thanks-to: Giuseppe D'Angelo

+++ plugin/pango/gvtextlayout_pango.c
@@ -100,7 +100,7 @@ static bool pango_textlayout(textspan_t * span, char **fontpath)
         !is_exactly_equal(fontsize, span->font->size)) {
 
 	/* check if the conversion to Pango units below will overflow */
-	if (INT_MAX / PANGO_SCALE < span->font->size) {
+	if ((double)INT_MAX / PANGO_SCALE < span->font->size) {
 	    return false;
 	}
